pyproject.toml
[project]
name = "foji-worker"
version = "0.1.0"
description = "Foji AI — Background processing Lambda workers"
requires-python = ">=3.12"
dependencies = [
    "sqlalchemy[asyncio]>=2.0.36",
    "psycopg2-binary>=2.9.10",
    "pydantic-settings>=2.6.0",
    "boto3>=1.35.0",
    "pdfplumber>=0.11.0",
    "python-docx>=1.1.2",
    "python-pptx>=1.0.2",
    "openpyxl>=3.1.5",
    "tiktoken>=0.8.0",
    "httpx>=0.27.0",
    "openai>=1.57.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "ruff>=0.8.0",
]

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.backends.legacy:build"

[tool.ruff]
line-length = 100
target-version = "py312"
